vcpkg integration
Overlay ports for projects (sleigh, lief) that don't exist in upstream
vcpkg repo and/or are required overlays (mbedtls v3) because upstream
won't accept them due to single-version package policy and mandatory
compatibility of all existing packages.

This uses a development version of sleigh that does a large refactor of
the CMake and some target names.

The lief portfile was copied from an existing PR